add events.{c,h} with functions to get/filter/dispatch events
Events that should be handled unconditionally (that is, independent of
the currently active mode, like dragging an icon), as if they are
handled by a background process have been abstracted out of the main
loop into a new file, event.c.

All event loops (main event or sub-events) now call get_drag_event() or
get_next_event().
